Update patch set 1
Patch Set 1:

treltexx is not receiving cm-12.1 nightlies.
I know people are building cm-12.1 off these branches though, so I figured it'd be nice for them to have this as well.
Also, even the cm-12.1 sepolicy is far from done for this device, so without this, the branch wouldn't even boot I think.

Patch-set: 1